refactor image-sampling into filter module
